[add] reorder, change user_name unique

+++ controllers/api_type.go
@@ -1,13 +1,13 @@
 package controllers
 
 type responseUserInfo struct {
-	Id       int    `json:"id"`
+	Id       uint   `json:"id"`
 	UserName string `json:"user_name"`
 	Email    string `json:"email"`
 }
 
 type createNewProjectBody struct {
-	ID          int64  `json:"id"`
+	ID          uint   `json:"id"`
 	ProjectName string `json:"project_name"`
 	Description string `json:"description"`
 	PointMan    string `json:"point_man"`
